Blog, BlogPost, etc.: refactors blog related data, images, and components into blog/ and updates references accordingly
