+++ src/data/blog-posts.json
@@ -12,7 +12,7 @@
         "subtitle": "Subtitle 2"
       },
       "image": {
-        "src": "/src/img/social-media-preview.png",
+        "src": "./img/social-media-preview.png",
         "alt": "alt-text"
       }
     },
@@ -25,7 +25,7 @@
       },
       {
         "image": {
-          "src": "/src/img/social-media-preview.png",
+          "src": "./img/social-media-preview.png",
           "alt": "alt-text",
           "width": "100%",
           "caption": "Social media preview"
